Moved yeast parsing into idg2sl/idg2sl_parser_yeast.py
