+++ idg2sl/__init__.py
@@ -1,6 +1,7 @@
 from .synthetic_lethal_interaction import SyntheticLethalInteraction
 from .entrez_parser import EntrezParser
 from .idg2sl_parser_human import *
+from .idg2sl_parser_yeast import *
 
 
 __all__ = ["SyntheticLethalInteraction", "EntrezParser"]
